backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
django-filter>=23.0
django-cors-headers>=4.0
pandas>=2.0
openpyxl>=3.1
gunicorn>=21.0
whitenoise>=6.6
